implemented the shooting ships

+++ src/gui.h
@@ -21,7 +21,7 @@ struct HUDItems {
   unsigned int bulletCount;
   float altitude;
   unsigned int score;
-  float elapsedTime;
+  double elapsedTime;
   bool crashed;
 };
